+++ test/utOrbit.py
@@ -8,6 +8,7 @@ import sys, os
 import numpy as np
 import random
 
+machine_initialized = False
 
 class TestOrbit(unittest.TestCase):
     """
@@ -15,16 +16,25 @@ class TestOrbit(unittest.TestCase):
 
     - orbit dimension
     """
-    _eget = 'default.eget'
-    _eput = 'default.eput'
 
     def setUp(self):
-        self.assertTrue(os.path.exists(CFAPKL))
-        self.orbit = hla.orbit.Orbit(hla._cfa)
-        pass
+        global machine_initialized
+        if not machine_initialized:
+            hla.machines.initNSLS2VSR()
+            hla.machines.initNSLS2VSRTxt()
+            machine_initialized = True
+        self.lat = hla.machines.getLattice('SR')
+        self.assertTrue(self.lat)
 
     def test_orbit(self):
-        self.assertTrue(len(self.orbit.get()) > 0)
+        self.assertTrue(len(hla.getElements('BPM')) > 0)
+
+        v = hla.getOrbit()
+        self.assertTrue(len(v) > 0)
+        v = hla.getOrbit('*')
+        self.assertTrue(len(v) > 0)
+        v = hla.getOrbit('P*')
+        self.assertTrue(len(v) > 0)
 
 if __name__ == "__main__":
     unittest.main()